Added methods for choosing default reduction
Added default_reduction, which returns the default reduction of the object passed (#31). For Seurat objects, this method calls the existing SeuratObject method DefaultDimReduc. For SingleCellExperiment objects, the method searches the object for either a UMAP, a T-SNE, or a PCA projection, in that order, and returns the first one found. The reduction priority of UMAP, T-SNE, then PCA is consistent with the behavior of the Seurat DefaultDimReduc method.