Speed testing DimPlot function
Added section to drafts Rmd with a speed test comparing the performance of the current DimPlot function to the previous SCEPlots:DimPlot function, and to Seurat::DimPlot. The section includes a function to test the speed of any function over n trials, as well as the code for the previous DimPlot implementation. The speed of FetchData compared to directly accessing reduction data is also compared.