Test speed and size vs HDF5Array compression level
The speed of data access for SingleCellExperiment objects saved via `HDF5Array::saveHDF5SummarizedExperiment` depends on the compression level chosen when saving to disk. Lower compression levels require more disk space, but allow data to be retrieved more quickly. The speed of data access was assessed for each compression level chosen, and for the default compression level, and the size on disk was also recorded.