Fix error with labeling feature plot
When setting label = TRUE on plot_feature, an error was observed with Seurat::LabelClusters during plot creation. The cause was the label_by column not being a factor. In Seurat::FeaturePlot, Ident() is passed to LabelClusters instead of label_by metadata, and this is always a factor. Seurat::LabelClusters takes the Ident column and operates on it as if it is a factor, and if a character vector is passed instead, the function will convert the data to NAs and return an error.

The issue was fixed by converting the label_by metadata column to a factor after it is fetched, if it is not one already.