S3 methods to fetch keys, determine key type
Seurat::ExIPlot uses Seurat object specific code to change labels on the x and y axes based on whether the feature entered is from an assay or a reduction. To adapt this code for SingleCellExperiment objects, two S3 methods were created (#42). `all_keys` returns all the valid "keys" for assays and reductions, and `key_type` reports if a key corresponds to an assay or a reduction.