fetch_cells method
S3 Method created to address #45. fetch_cells will return the cell names corresponding to specified levels of a metadata variable. This method currently only works with categorical metadata, and is not yet capable of cell selection via more complex conditional statements. Capabilities will be added to this method later on if needed.