fetch_metadata: default value of cells
fetch_metadata originally requires the `cells` parameter to be defined. In most use cases the user requests metadata for all the cells in an object, so it is sensible to have the function pull all cells by default. This has been implemented by adding a default value of NULL for cells, which will cause the method to run the get_all_cells() method on the object.