Determine best approach for pulling full metadata table
Tested pulling the full metadata table (via colData or @meta.data) compared to calling fetch_metadata with all the metadata variables selected. It was determined that pulling the full metadata directly was faster than calling fetch_metadata with all variable names specified, confirming that either a new function to perform this operation, or a modification to fetch_metadata that pulls the full table, is necessary.