Update handling of single variables by fetch_anndata
Converting vars to a list before passing it to fetch_anndata causes errors for multiple features. Single variable entries are converted to a string in fetch_anndata, which results in for loops in the function looping through each letter of the feature instead of looping once for the feature. This case is now handled in fetch_anndata by converting single-length variable entries to a list. Only entries that evaluate to a string data type in Python are converted.