Method for expr_plot labeling function
The expr_plot function for plotting violin and scatter plots contained code that did not work well for Anndata objects. The original code extracted a key from each feature entered and determined if the key was an assay or a reduction, changing the axis labels based on the type of feature. Since Anndata objects do not have enough structure to determine if a feature is a reduction based on the feature name, the behavior of expr_plot needed to be modified for Anndata objects. A new method, relabel_axis, was created to return the text to use for the axis label, and for Anndata objects the method will cause the label to be removed in all cases except for the case where "X_" appears in the feature name. Cases where features have multiple underscores, like "X_umap_1" result in the label being removed.

The key_type method originally used is no longer necessary, and has been deleted.