Fix error with dot plots for anndata objects
When plotting dot plots with anndata objects, dot plots returned a single column of dots with the label NA, instead of the expected behavior of one column of dots per feature. The cause was code involved in the generation of an intermediate table with summary statistics for the dot plot. One of the columns of the table was a factor with levels based on the names of the features entered by the user, but with values equal to the feature names returned from FetchData. For FetchData on Python objects, the modality/location key is always added to features with data pulled, regardless of whether the user entered the feature with a key. If there is a discrepency between the feature names entered and the feature names returned from FetchData, the factor produced in the dot plot will return NA values for that feature.

The issue was corrected by changing the levels of the factor to be equal to the column names of the table returned from FetchData instead of the names of the features entered.