Add default values of cells/dims to fetch_reduction
It is possible to call fetch_reduction without defining `cells` and `dims` without getting an error up front due to these parameters being undefined. Downstream computation for Seurat and SCE objects proceeds properly without specifying cells, but an error is returned for Anndata objects when the FetchData call downstream is called. Defaults are added for cells and dims to avoid these errors.