Fix bug with FetchData, plot_feature
When calling plot_feature on anndata objects with subset (via `cells`), an error was returned: a mismatch was observed in the number of rows in the reduction cordinates returned via fetch_reduction, and the expression data returned via FetchData. The cause of the error was found to be in the python fetch_anndata script: when pulling data from sparse matrices, the function always pulled data from the full object, instead of slicing for the cells passed to the function. This has been corrected here.