Fix #71
The issue described in #71 was caused by an attempt to set the column names of a matrix with no columns, in the event that all features passed to the function do not exist in the object. This has been resolved by adding a conditional statement that checks if the length of features found is at least equal to one before setting column names.