Errors for unsupported uses of fetch_cells
- Return an error when meta_levels is set to NULL
- Return an error when a numeric variable is passed to meta_var

fetch_cells only works for categorical variables, when a subset of the levels in a metadata variable is desired. The function will return a zero length character vector when numeric metadata is entered, which can complicate downstream operations. Unexpected behavior is prevented by returning an error in unsupported use cases.