Error handling for plot_reduction
Plot reduction did not previously return errors when numeric metadata was passed to the group_by variable, which resulted in the function attempting to create groups for each value of numeric metadata when a numeric variable was passed. To avoid the performance issues that result from this situation, code was added to return errors when numeric metadata is passed to group_by. The same was done for split_by and shape_by.