Error handling code for fetch_metadata
When a variable that doesn't exist in object metadata is passed to fetch_metadata, no error is returned, and for anndata objects NULL is returned for the table retrieved, which results in a cryptic error. To prevent unintended behavior, code was added to throw an informative error message in this situation.