Add update_object_metadata method from scExploreR
Added the update_object_metadata method. This was originally in  scExploreR, but it has now been determined that there is a use case for the function beyond the app. The function is useful for programmatically saving a metadata table that has been modified separate from the object (which currently happens in the plotting functions in scPlots), but this function could be problematic if used interactively in data analysis. To prevent misuse of the function, it will remain hidden from the user (by putting @noRd in the Roxygen documentation).