Fix Improper searching of reduction matrices (#90)
When there is an object with reduction names sharing the same root (i.e. "X_mofa" and "X_mofa_umap"), and a reduction coordinate variable such as X_mofa_umap_1 is entered, the fetch_keyed_vars function in fetch_anndata.py will search for X_mofa_umap_1 in both the X_mofa and the X_mofa_umap matrices. fetch_keyed_vars will remove the "key" of the matrix from the variable name before searching, meaning it will search in "X_mofa" for "umap_1", and it will search in "X_mofa_umap" for "1". Only the latter is intended, and the error in #90 ocurrs when the former variable is being searched for.

To fix this issue, I added error handling to the fetch_keyed_vars function, directing it to continue searching after requesting a non-existent feature, and return no data from that matrix. This would result in data not being returned properly if a user searched for coordinates from "X_mofa" and "X_mofa_umap" concurrently, but this does not seem likely.