Remove plotting functions from pkgdown website
The plotting functions in SCUBA are beyond the scope of the package. They should be retained in SCUBA (for now), but they should not be on the website. To preserve docs and to keep users from having to call `:::` to access these functions (which would break scExploreR and other downstream scripts), `@keyword internal` was added to the roxygen docs for all plotting functions.