Remove scExploreR specific behavior from features_in_assay
The methods in features_in_assay depended on the assay_names generic, which is defined in the context of scExploreR. This generic depends on an anndata property that is only present in objects prepared for scExploreR (uns.scExploreR_assays), so this method shouldn't be copied. Instead, checks in the features_in_assay function were modified to fit SCUBA-specific usage contexts.