Added fetch_data generic
The `fetch_data` generic has been added, to allow SCUBA to control the generic used for the FetchData methods. This was done to eliminate the need to load the Seurat package to run SCUBA's FetchData methods created for other packages, and to allow for consistent case to be used for all generics and functions in SCUBA.

This generic explicitly executes the SeuratObject FetchData method when a Seurat object is passed. For all other objects, it runs the FetchData methods by SCUBA. These were transferred unchanged from `R/FetchData.R`.

The FetchData methods have not been removed. They will be deprecated in the next version.