Hide FetchData methods
Usage of the FetchData methods is discouraged since SCUBA does not own the generic for these. Now that fetch_data exists, these should be deprecated in favor of the fetch_data generic. The FetchData methods have been hidden with keywords internal to keep them from appearing on the pkgdown website and the package documentation page.