Manage return class of features_in_assay
For MuData objects, var_names can be a Pandas Index. In this case, it is not properly converted to a character vector by Reticulate upon return. If var_names is a Pandas Index, it is now converted to a list in Python first, so it can be properly returned as a character vector in R.