Fix bug with fetch_data, match_vars
fetch_keyed_vars_mudata calls match_vars to construct a dictionary with the locations of each variable requested. This function creates a special element, "mod_obs", which points to variables that are metadata to be pulled from a specific modality. If no vars match mod_obs, the "mod_obs" element of the dictionary returned will be empty, which will cause fetch_keyed_vars to fetch metadata for nothing, and append the result to the exisitng vars from fetch_mudata, causing an error. Now, if no variables match a mod_obs key, the "mod_obs" element will not be created.