Remove unnecessary step from testing action
When managing environments via uv, it is no longer necessary to create a virtual environment for Reticulate. Furthermore, if RETICULATE_PYTHON is set, the ephemeral environment managed by uv will be ignored when running Reticulate. The virtual environment step should be removed now that the Python environment is being managed in uv via py_require.